[#131] feat:프로필설정페이지 닉네임만 서버전송 성공시 에러 로직 수정
